Add common issue templates for all Stellar repos (#5)
Add common issue templates for all Stellar repos using simple language and concepts that are accessible and inclusive. Also add a link when creating a new issue for asking questions to direct people to our StackExchange.

Why:
@ire-and-curses and I were talking this week about how we receive bug reports on some repositories that often provide no description. Asking some simple questions upfront, and asking kindly for people to answer those questions, goes a long way to having people provide enough details such that their bug report becomes a great contribution.

I see that we have some solid issue templates already on the `stellar/js-stellar-sdk` project, so I copied them as a starting point.

The issue templates in this repository request the same information, but reword the questions into specific single asks using the simplest English language terms and concepts I could find. This is in a similar fashion to how we have simplified language in the `stellar/go`s PR template to make it more inclusive and accessible to a wider variety of people. Some of the questions I reworded I used as inspiration the issue template of the golang/go project.

In the bug report template a request for providing an example of the bug in an online editor has been added, because this will greatly assist in us reproducing a problem and make the process of reviewing incoming bugs more efficient. This process is used by the golang/go project with a lot of success. I also see examples in the js-stellar-sdk project where folks have reported issues, and we've then posted jsfiddle.net examples to confirm the exact problem, so it'd be great to turn that around and encourage others to submit the example breaking code.

The link to our StackExchange will signal boost and grow that community.