Add Stellar logo (#9)
### What
Add Stellar logo.

### Why
Several of our repositories have a Stellar logo at the top of them and that image is now broken.

https://github.com/stellar/stellar-protocol
https://github.com/stellar/stellar-core
https://github.com/stellar/go
https://github.com/stellar/js-stellar-sdk

The image was being hosted by stellar.org but with the recent new website launch the logo is no longer hosted. Given that the website can change over time we should host the logo for the GitHub projects on GitHub. I think we should be able to host it here in this repository in one place and link to it from the projects that display it.

I used archive.org to find the original image that we were using:
https://web.archive.org/web/20191224202119if_/https://www.stellar.org/old-content/2019/03/stellar-logo-solo-1.png