Add link to stellar-dev mailing list to issues page of our repos (#10)
### What
Add link to stellar-dev mailing list to issues page of our repos.

### Why
We don't point people to the stellar-dev mailing list on our repos but
we want people to use the mailing list for discussion. We should promote
it front and center when people are thinking of communicating with us on
GitHub. In some ways I think it is more important than the link to the
stellar-protocol repo that we have in this list.