Add profile readme (#15)
Add a profile readme that will display on github.com/stellar.

GitHub released the ability for us to create profile readmes. Here is an example of one that Twitter is using: github.com/twitter. It can be helpful to link from the profile readme to all the typical places we try to direct folks like the mailing list, docs, where to ask questions. We link to most of these places on the "New Issue" screens of repos to increase discoverability of them, and this would be a good place too.

Most of the copy I copied from stellar.org and I imagine other folks may wish to improve it.